requirements.txt
pytest
pytest-cov
atproto
streamlit
pytrends
sqlalchemy
email_validator
streamlit_agraph
pandas>=1.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
vaderSentiment>=3.3.2
certifi>=2023.5.7
flask>=2.2.0
boto3>=1.28.0
aioboto3>=11.0.0
freezegun>=1.2.0
requests
scikit-learn
numpy